Fully automated from .roh file

+++ Rohan/sniptype.h
@@ -12,7 +12,7 @@ typedef enum {Error_ = -1, Success_, False_ = 0, True_} Boolean_T;
 
 #if defined(__unix__)
  typedef unsigned char  BYTE;
- typedef unsigned long  DWORD;
+ typedef unsigned int  DWORD;
  typedef unsigned short WORD;
  #if !defined(FAR)
   #define FAR
@@ -36,7 +36,7 @@ typedef enum {Error_ = -1, Success_, False_ = 0, True_} Boolean_T;
       && !defined(_WINDOWS)
  #if !defined(OS2)
   typedef unsigned char  BYTE;
-  typedef unsigned long  DWORD;
+  typedef unsigned int  DWORD;
  #endif
  typedef unsigned short WORD;
 #else
@@ -62,7 +62,7 @@ typedef union {
 
 typedef union {
       VAR16_            v16[2];
-      signed long       l;
+      signed int       l;
       DWORD             dw;
       float             f;
       void              *p;
